Move islamic code to calendrical_calculations
