+++ utils/calendrical_calculations/src/iso.rs
@@ -43,7 +43,7 @@ pub fn fixed_from_iso(year: i32, month: u8, day: u8) -> RataDie {
 }
 
 // Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1191-L1217
-fn iso_year_from_fixed(date: RataDie) -> i64 {
+pub(crate) fn iso_year_from_fixed(date: RataDie) -> i64 {
     // Shouldn't overflow because it's not possbile to construct extreme values of RataDie
     let date = date - EPOCH;
