+++ utils/calendrical_calculations/src/lib.rs
@@ -35,6 +35,7 @@ pub mod error;
 pub mod ethiopian;
 pub mod hebrew;
 pub mod helpers;
+pub mod islamic;
 pub mod iso;
 pub mod julian;
 pub mod persian;
